Refactor visuals module: add components and plugin structure; clean up unused code in board components
