+++ game/src/main.rs
@@ -1,5 +1,6 @@
 mod app;
 mod board;
+mod visuals;
 
 fn main() {
     app::run();
