+++ game/src/visuals/mod.rs
@@ -1,1 +1,5 @@
+pub mod components;
+mod plugin;
+pub mod systems;
 
+pub use plugin::VisualsPlugin;
