+++ game/src/visuals/plugin.rs
@@ -1,1 +1,9 @@
+use bevy::prelude::*;
 
+pub struct VisualsPlugin;
+
+impl Plugin for VisualsPlugin {
+    fn build(&self, app: &mut App) {
+        // TODO make systems and add them
+    }
+}
